Add files via upload
Latest development before discontinuing this project in April 2025 due to no longer staying involved with the nerf community. 
This worked in concept, and the code did what I wanted it to do. This was never implemented into a final design, but now I have the code and diagrams to rebuild should I ever want to revisit.